refactor(add): split command flow helpers
